[IMP] payroll: Remove hr_contract_advantages from payroll module
[IMP] payroll: remove old model from adv_template file

[IMP] payroll: Remove hr_contract_advantages from payroll module
