+++ payroll/__manifest__.py
@@ -19,7 +19,6 @@
         "wizard/hr_payroll_payslips_by_employees_views.xml",
         "views/menus.xml",
         "views/hr_contract_views.xml",
-        "views/hr_contract_advantage_views.xml",
         "views/hr_payroll_structure_views.xml",
         "views/hr_salary_rule_category_views.xml",
         "views/hr_contribution_register_views.xml",
