+++ payroll/models/__init__.py
@@ -1,8 +1,6 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import hr_contract
-from . import hr_contract_advantage_template
-from . import hr_contract_advantage
 from . import hr_employee
 from . import hr_payroll_structure
 from . import hr_salary_rule
