+++ payroll_advantages/hr_contract_advantage_views.xml
